gradle fix

+++ src/main/java/io/github/kydzombie/cairn/api/packet/UpdatePacketHelper.java
@@ -22,7 +22,7 @@ public class UpdatePacketHelper {
     private static final Map<Class<?>, Serializer<?>> SERIALIZERS = new HashMap<>();
 
     static {
-        registerSerializer(Boolean.class, (buffer, value) -> buffer.put(value ? (byte) 1 : (byte) 0), (buffer) -> buffer.get() > 0, (value) -> 1);
+        registerSerializer(Boolean.class, (buffer, value) -> buffer.put((byte) (value ? 1 : 0)), (buffer) -> buffer.get() > 0, (value) -> 1);
         registerSerializer(Byte.class, ByteBuffer::put, ByteBuffer::get, (value) -> 1);
         registerSerializer(Short.class, ByteBuffer::putShort, ByteBuffer::getShort, (value) -> 2);
         registerSerializer(Integer.class, ByteBuffer::putInt, ByteBuffer::getInt, (value) -> 4);
@@ -81,7 +81,8 @@ public class UpdatePacketHelper {
                 nbt.write(dataOut);
 
                 return byteOut.size() + 4;
-            });
+            }
+        );
     }
 
     public static <T> void registerSerializer(Class<T> type, BiConsumer<ByteBuffer, T> serializer, Function<ByteBuffer, T> deserializer, Function<T, Integer> sizeFunc) {
